# docs/swagger.yaml
definitions:
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
  dto.SaveFlagRequest:
    properties:
      description:
        type: string
      environment:
        type: string
      name:
        type: string
      status:
        type: string
    type: object
  dto.UpdateFlagStatusRequest:
    properties:
      status:
        type: string
    type: object
info:
  contact: {}
  description: API для управления фича-флагами
  title: Feature Flags API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Возвращает JWT токен
      parameters:
      - description: Данные для входа
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses: {}
      summary: Вход в систему
      tags:
      - auth
  /flags:
    get:
      description: Возвращает массив фича-флагов с возможностью фильтрации
      parameters:
      - description: Поиск по имени
        in: query
        name: search
        type: string
      - description: Фильтр по окружению (dev, staging, production)
        in: query
        name: environment
        type: string
      - description: Фильтр по статусу (enabled, disabled)
        in: query
        name: status
        type: string
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Получить список всех флагов
      tags:
      - flags
    post:
      consumes:
      - application/json
      description: Создает новый фича-флаг и привязывает его к создателю и команде
      parameters:
      - description: Данные нового флага
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SaveFlagRequest'
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Создать новый флаг
      tags:
      - flags
  /flags/{id}:
    get:
      description: Возвращает фич-флаг по его идентификатору
      parameters:
      - description: ID флага
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Получить фич-флаг по id
      tags:
      - flags
    put:
      consumes:
      - application/json
      description: Редактирует уже существующий фича-флаг
      parameters:
      - description: ID флага
        in: path
        name: id
        required: true
        type: integer
      - description: Данные нового флага
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SaveFlagRequest'
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Редактировать фича-флаг
      tags:
      - flags
  /flags/{id}/status:
    patch:
      consumes:
      - application/json
      description: Редактирует уже существующий фича-флаг
      parameters:
      - description: ID флага
        in: path
        name: id
        required: true
        type: integer
      - description: Новый статус
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateFlagStatusRequest'
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Изменить статус фича-флаг
      tags:
      - flags
  /me:
    get:
      description: Возвращает информацию о пользователе по токену из заголовка Authorization
      produces:
      - application/json
      responses: {}
      security:
      - ApiKeyAuth: []
      summary: Получить данные текущего пользователя
      tags:
      - auth
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"